requirements/base.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    ./bin/compile_dependencies.sh
#
amqp==5.2.0
    # via kombu
ape-pie==0.2.0
    # via zgw-consumers
asgiref==3.8.1
    # via
    #   django
    #   django-axes
    #   django-cors-headers
asn1crypto==1.5.1
    # via webauthn
attrs==23.2.0
    # via
    #   glom
    #   jsonschema
    #   referencing
billiard==4.2.0
    # via celery
bleach==6.1.0
    # via open-api-framework
boltons==24.0.0
    # via
    #   face
    #   glom
cbor2==5.6.4
    # via webauthn
celery==5.4.0
    # via
    #   flower
    #   notifications-api-common
    #   open-api-framework
certifi==2024.6.2
    # via
    #   elastic-apm
    #   requests
    #   sentry-sdk
cffi==1.16.0
    # via cryptography
charset-normalizer==3.3.2
    # via requests
click==8.1.7
    # via
    #   celery
    #   click-didyoumean
    #   click-plugins
    #   click-repl
click-didyoumean==0.3.1
    # via celery
click-plugins==1.1.1
    # via celery
click-repl==0.3.0
    # via celery
commonground-api-common==1.13.1
    # via open-api-framework
coreapi==2.3.3
    # via commonground-api-common
coreschema==0.0.4
    # via coreapi
cryptography==42.0.8
    # via
    #   django-simple-certmanager
    #   josepy
    #   mozilla-django-oidc
    #   pyopenssl
    #   webauthn
django==4.2.13
    # via
    #   commonground-api-common
    #   django-admin-index
    #   django-appconf
    #   django-axes
    #   django-cors-headers
    #   django-filter
    #   django-formtools
    #   django-jsonform
    #   django-log-outgoing-requests
    #   django-markup
    #   django-otp
    #   django-phonenumber-field
    #   django-privates
    #   django-redis
    #   django-relativedelta
    #   django-rest-framework-condition
    #   django-sendfile2
    #   django-setup-configuration
    #   django-simple-certmanager
    #   django-solo
    #   django-treebeard
    #   django-two-factor-auth
    #   djangorestframework
    #   djangorestframework-inclusions
    #   drf-nested-routers
    #   drf-spectacular
    #   drf-yasg
    #   maykin-2fa
    #   mozilla-django-oidc
    #   mozilla-django-oidc-db
    #   notifications-api-common
    #   open-api-framework
    #   zgw-consumers
django-admin-index==3.1.1
    # via open-api-framework
django-appconf==1.0.6
    # via django-log-outgoing-requests
django-axes==6.5.1
    # via open-api-framework
django-cors-headers==4.4.0
    # via open-api-framework
django-filter==24.2
    # via
    #   commonground-api-common
    #   open-api-framework
django-formtools==2.5.1
    # via django-two-factor-auth
django-jsonform==2.22.0
    # via
    #   mozilla-django-oidc-db
    #   open-api-framework
django-log-outgoing-requests==0.6.1
    # via open-api-framework
django-markup==1.8.1
    # via open-api-framework
django-ordered-model==3.7.4
    # via django-admin-index
django-otp==1.5.0
    # via django-two-factor-auth
django-phonenumber-field==7.3.0
    # via django-two-factor-auth
django-privates==2.0.0.post1
    # via django-simple-certmanager
django-redis==5.4.0
    # via open-api-framework
django-relativedelta==2.0.0
    # via zgw-consumers
django-rest-framework-condition==0.1.1
    # via commonground-api-common
django-sendfile2==0.7.1
    # via django-privates
django-setup-configuration==0.1.0
    # via open-api-framework
django-simple-certmanager==2.0.0
    # via zgw-consumers
django-solo==2.3.0
    # via
    #   commonground-api-common
    #   django-log-outgoing-requests
    #   mozilla-django-oidc-db
    #   notifications-api-common
    #   zgw-consumers
django-treebeard==4.7.1
    # via -r requirements/base.in
django-two-factor-auth[phonenumberslite,webauthn]==1.16.0
    # via maykin-2fa
djangorestframework==3.15.2
    # via
    #   commonground-api-common
    #   djangorestframework-gis
    #   djangorestframework-inclusions
    #   drf-nested-routers
    #   drf-spectacular
    #   drf-yasg
    #   notifications-api-common
    #   open-api-framework
djangorestframework-camel-case==1.4.2
    # via
    #   commonground-api-common
    #   notifications-api-common
djangorestframework-gis==1.0
    # via open-api-framework
djangorestframework-inclusions==1.2.0
    # via open-api-framework
drf-nested-routers==0.94.1
    # via commonground-api-common
drf-spectacular==0.27.2
    # via open-api-framework
drf-yasg==1.21.7
    # via commonground-api-common
ecs-logging==2.2.0
    # via elastic-apm
elastic-apm==6.22.3
    # via open-api-framework
face==20.1.1
    # via glom
flower==2.0.1
    # via open-api-framework
furl==2.1.3
    # via ape-pie
gemma-zds-client==2.0.0
    # via
    #   commonground-api-common
    #   notifications-api-common
glom==23.5.0
    # via mozilla-django-oidc-db
humanize==4.9.0
    # via flower
idna==3.7
    # via requests
inflection==0.5.1
    # via
    #   drf-spectacular
    #   drf-yasg
iso-639==0.4.5
    # via commonground-api-common
isodate==0.6.1
    # via commonground-api-common
itypes==1.2.0
    # via coreapi
jinja2==3.1.4
    # via coreschema
josepy==1.14.0
    # via mozilla-django-oidc
jsonschema==4.22.0
    # via drf-spectacular
jsonschema-specifications==2023.12.1
    # via jsonschema
kombu==5.3.7
    # via celery
markupsafe==2.1.5
    # via jinja2
maykin-2fa==1.0.0
    # via open-api-framework
mozilla-django-oidc==4.0.1
    # via mozilla-django-oidc-db
mozilla-django-oidc-db==0.19.0
    # via open-api-framework
notifications-api-common==0.2.2
    # via commonground-api-common
open-api-framework==0.6.0
    # via -r requirements/base.in
orderedmultidict==1.0.1
    # via furl
oyaml==1.0
    # via commonground-api-common
packaging==24.1
    # via drf-yasg
phonenumberslite==8.13.40
    # via django-two-factor-auth
pillow==10.4.0
    # via -r requirements/base.in
prometheus-client==0.20.0
    # via flower
prompt-toolkit==3.0.47
    # via click-repl
psycopg2==2.9.9
    # via open-api-framework
pycparser==2.22
    # via cffi
pyjwt==2.8.0
    # via
    #   commonground-api-common
    #   gemma-zds-client
    #   zgw-consumers
pyopenssl==24.1.0
    # via
    #   josepy
    #   webauthn
pypng==0.20220715.0
    # via qrcode
python-dateutil==2.9.0.post0
    # via
    #   celery
    #   django-relativedelta
python-decouple==3.8
    # via open-api-framework
python-dotenv==1.0.1
    # via open-api-framework
pytz==2024.1
    # via
    #   drf-yasg
    #   flower
pyyaml==6.0.1
    # via
    #   drf-spectacular
    #   drf-yasg
    #   gemma-zds-client
    #   oyaml
qrcode==7.4.2
    # via django-two-factor-auth
redis==5.0.7
    # via django-redis
referencing==0.35.1
    # via
    #   jsonschema
    #   jsonschema-specifications
requests==2.32.3
    # via
    #   ape-pie
    #   commonground-api-common
    #   coreapi
    #   django-log-outgoing-requests
    #   gemma-zds-client
    #   mozilla-django-oidc
    #   open-api-framework
    #   zgw-consumers
rpds-py==0.18.1
    # via
    #   jsonschema
    #   referencing
sentry-sdk==2.7.1
    # via open-api-framework
six==1.16.0
    # via
    #   bleach
    #   furl
    #   isodate
    #   orderedmultidict
    #   python-dateutil
sqlparse==0.5.0
    # via django
tornado==6.4.1
    # via flower
typing-extensions==4.12.2
    # via
    #   mozilla-django-oidc-db
    #   qrcode
    #   zgw-consumers
tzdata==2024.1
    # via celery
uritemplate==4.1.1
    # via
    #   coreapi
    #   drf-spectacular
    #   drf-yasg
urllib3==2.2.2
    # via
    #   elastic-apm
    #   requests
    #   sentry-sdk
uwsgi==2.0.26
    # via open-api-framework
vine==5.1.0
    # via
    #   amqp
    #   celery
    #   kombu
wcwidth==0.2.13
    # via prompt-toolkit
webauthn==2.2.0
    # via django-two-factor-auth
webencodings==0.5.1
    # via bleach
wrapt==1.16.0
    # via elastic-apm
zgw-consumers==0.33.0
    # via
    #   notifications-api-common
    #   open-api-framework
